--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -63,6 +63,8 @@ jobs:
           name: Run tests
           command: truffle test 
           working_directory: ~/repo/plasma_framework
+          environment:
+            MOCHA_REPORTER: eth-gas-reporter
 
   Truffle linter:
     executor: truffle_executor