--- a/plasma_framework/codechecks.yml
+++ b/plasma_framework/codechecks.yml
@@ -1,0 +1,2 @@
+checks:
+  - name: eth-gas-reporter/codechecks