--- a/plasma_framework/package.json
+++ b/plasma_framework/package.json
@@ -13,6 +13,7 @@
     "eslint": "^5.3.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.17.3",
+    "eth-gas-reporter": "^0.2.6",
     "eth-sig-util": "^2.2.0",
     "ethereumjs-util": "^6.1.0",
     "openzeppelin-test-helpers": "^0.4.0",