--- a/plasma_framework/package.json
+++ b/plasma_framework/package.json
@@ -6,7 +6,8 @@
     "test": "test"
   },
   "dependencies": {
-    "openzeppelin-solidity": "^2.3.0"
+    "openzeppelin-solidity": "^2.3.0",
+    "dotenv": "^8.0.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",