--- a/plasma_framework/package.json
+++ b/plasma_framework/package.json
@@ -10,6 +10,7 @@
     "dotenv": "^8.0.0"
   },
   "devDependencies": {
+    "@codechecks/client": "^0.1.5",
     "chai": "^4.2.0",
     "eslint": "^5.3.0",
     "eslint-config-airbnb-base": "^13.1.0",