### Note
- adds eth-gas-reporter to help us profile our gas usage.
- https://www.npmjs.com/package/eth-gas-reporter
- I do find a down side of setting this, the unit test speed become slower : (
  - without reporter: truffle test  42.05s user 3.95s system 52% cpu 1:26.89 total
  - with reporter: truffle test  72.68s user 43.38s system 60% cpu 3:11.32 total
  - thus 2~3 times slower for current tests.


### Test

```
·------------------------------------------------------------------|----------------------------|-------------|----------------------------·
|               Solc version: 0.5.8+commit.23d335f2                ·  Optimizer enabled: false  ·  Runs: 200  ·  Block limit: 6721975 gas  │
···································································|····························|·············|·····························
|  Methods                                                         ·                1 gwei/gas                ·       225.24 usd/eth       │
·····································|·····························|··············|·············|·············|··············|··············
|  Contract                          ·  Method                     ·  Min         ·  Max        ·  Avg        ·  # calls     ·  usd (avg)  │
·····································|·····························|··············|·············|·············|··············|··············
|  BlockController                   ·  registerVault              ·       65896  ·      66005  ·      65990  ·          46  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  BlockController                   ·  submitBlock                ·           -  ·          -  ·      75752  ·           8  ·       0.02  │
·····································|·····························|··············|·············|·············|··············|··············
|  BlockController                   ·  submitDepositBlock         ·           -  ·          -  ·      72433  ·          11  ·       0.02  │
·····································|·····························|··············|·············|·············|··············|··············
|  Erc20Vault                        ·  deposit                    ·      118186  ·     122414  ·     121568  ·           5  ·       0.03  │
·····································|·····························|··············|·············|·············|··············|··············
|  Erc20Vault                        ·  setDepositVerifier         ·           -  ·          -  ·      43429  ·          10  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  EthVault                          ·  deposit                    ·      110193  ·     110374  ·     110302  ·          10  ·       0.02  │
·····································|·····························|··············|·············|·············|··············|··············
|  EthVault                          ·  setDepositVerifier         ·           -  ·          -  ·      43429  ·          14  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  ExitGameController                ·  addToken                   ·      822048  ·     822112  ·     822109  ·          24  ·       0.19  │
·····································|·····························|··············|·············|·············|··············|··············
|  ExitGameController                ·  processExits               ·       32069  ·      44347  ·      41680  ·          15  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  ExitGameController                ·  registerExitGame           ·       65940  ·      66610  ·      66412  ·          38  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  OutputGuardParserRegistry         ·  registerOutputGuardParser  ·       44286  ·      44651  ·      44377  ·           4  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  OutputGuardParserRegistry         ·  renounceOwnership          ·           -  ·          -  ·      14385  ·           1  ·       0.00  │
·····································|·····························|··············|·············|·············|··············|··············
|  PaymentExitGame                   ·  registerSpendingCondition  ·       45338  ·      45466  ·      45402  ·          13  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  PaymentExitGame                   ·  startStandardExit          ·      227107  ·     320013  ·     267677  ·          21  ·       0.06  │
·····································|·····························|··············|·············|·············|··············|··············
|  PaymentSpendingConditionRegistry  ·  registerSpendingCondition  ·       44618  ·      44682  ·      44656  ·           5  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  PaymentSpendingConditionRegistry  ·  renounceOwnership          ·           -  ·          -  ·      14385  ·           1  ·       0.00  │
·····································|·····························|··············|·············|·············|··············|··············
|  PlasmaFramework                   ·  registerExitGame           ·           -  ·          -  ·      66566  ·           6  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  PlasmaFramework                   ·  registerVault              ·           -  ·          -  ·      66567  ·           6  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  PlasmaFramework                   ·  submitBlock                ·           -  ·          -  ·      76227  ·           4  ·       0.02  │
·····································|·····························|··············|·············|·············|··············|··············
|  PriorityQueue                     ·  delMin                     ·       25752  ·      36503  ·      30360  ·           7  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  PriorityQueue                     ·  insert                     ·       54160  ·      69160  ·      64359  ·          20  ·       0.01  │
·····································|·····························|··············|·············|·············|··············|··············
|  Deployments                                                     ·                                          ·  % of limit  ·             │
···································································|··············|·············|·············|··············|··············
|  BlockController                                                 ·           -  ·          -  ·     926560  ·      13.8 %  ·       0.21  │
···································································|··············|·············|·············|··············|··············
|  Erc20DepositVerifier                                            ·           -  ·          -  ·    1402606  ·      20.9 %  ·       0.32  │
···································································|··············|·············|·············|··············|··············
|  Erc20Vault                                                      ·           -  ·          -  ·    1263777  ·      18.8 %  ·       0.28  │
···································································|··············|·············|·············|··············|··············
|  EthDepositVerifier                                              ·           -  ·          -  ·    1279989  ·      19.1 %  ·       0.29  │
···································································|··············|·············|·············|··············|··············
|  EthVault                                                        ·      932681  ·     932745  ·     932736  ·      13.9 %  ·       0.21  │
···································································|··············|·············|·············|··············|··············
|  ExitGameController                                              ·           -  ·          -  ·    4076297  ·      60.7 %  ·       0.92  │
···································································|··············|·············|·············|··············|··············
|  OutputGuardParserRegistry                                       ·           -  ·          -  ·     676737  ·      10.1 %  ·       0.15  │
···································································|··············|·············|·············|··············|··············
|  PaymentExitGame                                                 ·     4973766  ·    4973830  ·    4973819  ·        74 %  ·       1.12  │
···································································|··············|·············|·············|··············|··············
|  PaymentOutputToPaymentTxCondition                               ·     1912998  ·    1914278  ·    1913584  ·      28.5 %  ·       0.43  │
···································································|··············|·············|·············|··············|··············
|  PaymentSpendingConditionRegistry                                ·           -  ·          -  ·     714626  ·      10.6 %  ·       0.16  │
···································································|··············|·············|·············|··············|··············
|  PlasmaFramework                                                 ·     5000425  ·    5000489  ·    5000480  ·      74.4 %  ·       1.13  │
···································································|··············|·············|·············|··············|··············
|  PriorityQueue                                                   ·           -  ·          -  ·    1032935  ·      15.4 %  ·       0.23  │
·------------------------------------------------------------------|--------------|-------------|-------------|--------------|-------------·
```