@boolafish 2019-07-22
Question: how do people feel with this? I don't like the fact it slows down the test and honestly we don't need to profile every time. However, I do like the report and profiling feature (though I recognize some functions should be hit but the gas report is not shown :( )

We could:
1. Add on our test setup. So we would have more understanding on our gas data per change. However, it makes our local development process slightly slower as the tests run slower.
1. Don't do anything. Only use periodically when we want to profile our contract.

Ideally I'd like to only add this to our CI, however, I find it not trivial to do so. It rely on `mocha` config inside the file `truffle-config.js`, which I don't see anyway to pass in flag to disable the reporter from truffle. eg. `truffle test --reporter xxxx`. The only (really ugly) way I can think of is in CI we "write the config file on the fly". So the CI would write the `mocha` part into our `truffle-config.js`. Given it's such an ugly approach I don't like that too : (

cc. @paulperegud @kevsul @pgebal @pik694