@pik694 2019-07-28
One other approach is to have two configs and let the CI choose the appropriate one.
It would mean maintaining two configs, though.