@kevsul 2019-08-05
We can optionally set the reporter in CI using an environment variable. In truffle-config.js
```
    mocha: {
        reporter: process.env.MOCHA_REPORTER || '',
```

By default there's no reporter. 
But we can enable it in CI by setting the env `MOCHA_REPORTER=eth-gas-reporter`.