@boolafish 2019-08-05
@kevsul oh that's smart and clean!